Accept single-digit date
From: Joe Edmonds 

I'm using MissingH.Wash.Mail.MailParser to parse mail sent from my
PalmOS treo using Versamail (3.1B).  Versamail seems to set the hour
in the "Date" header to a single-digit date, in violation of RFC2822.

If you want to follow the "be strict in what you emit and permissive
in what you accept" philosophy, attached is a patch to accept this
format.

- Joe Edmonds