Make it build with ghc-6.6 and 6.8 using configurations
requires one slightly less than pleasent hack because Patternsignatures
is not recognised by ghc-6.6 or the cabal that comes with 6.8.1, only
by the latest cabal that comes with 6.8.2.
Drop GHC-Options: -O2 as Cabal adds optimizations by default but also
lets the user configure with --disable-optimization, so there is no longer
any need to force optimisations on all the time and it prevents users
configuring without optimisations for quick builds.